requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
idna==2.6
python-dateutil==2.6.1
requests==2.18.4
selenium==3.8.1
six==1.11.0
urllib3==1.22
